requirements.txt
flask>=2.3.3
google-generativeai>=0.4.1
pymongo>=4.6.1
numpy>=1.24.3
pandas>=2.0.3
python-dotenv>=1.0.0
requests>=2.31.0